refactor: Move `tox_new_log` to `auto_test_support`.
This is only used in auto tests.
