+++ auto_tests/auto_test_support.c
@@ -1,3 +1,4 @@
+#include <assert.h>  // assert
 #include <stdlib.h>  // calloc, free
 
 #include "check_compat.h"
@@ -322,4 +323,71 @@ void run_auto_test(struct Tox_Options *options, uint32_t tox_count, void test(Au
     }
 
     free(autotoxes);
+}
+
+static const char *tox_log_level_name(Tox_Log_Level level)
+{
+    switch (level) {
+        case TOX_LOG_LEVEL_TRACE:
+            return "TRACE";
+
+        case TOX_LOG_LEVEL_DEBUG:
+            return "DEBUG";
+
+        case TOX_LOG_LEVEL_INFO:
+            return "INFO";
+
+        case TOX_LOG_LEVEL_WARNING:
+            return "WARNING";
+
+        case TOX_LOG_LEVEL_ERROR:
+            return "ERROR";
+    }
+
+    return "<unknown>";
+}
+
+void print_debug_log(Tox *m, Tox_Log_Level level, const char *file, uint32_t line, const char *func,
+                     const char *message, void *user_data)
+{
+    if (level == TOX_LOG_LEVEL_TRACE) {
+        return;
+    }
+
+    uint32_t index = user_data ? *(uint32_t *)user_data : 0;
+    fprintf(stderr, "[#%u] %s %s:%u\t%s:\t%s\n", index, tox_log_level_name(level), file, line, func, message);
+
+    if (level == TOX_LOG_LEVEL_ERROR) {
+        fputs("Aborting test program\n", stderr);
+        abort();
+    }
+}
+
+Tox *tox_new_log_lan(struct Tox_Options *options, Tox_Err_New *err, void *log_user_data, bool lan_discovery)
+{
+    struct Tox_Options *log_options = options;
+
+    if (log_options == nullptr) {
+        log_options = tox_options_new(nullptr);
+    }
+
+    assert(log_options != nullptr);
+
+    tox_options_set_local_discovery_enabled(log_options, lan_discovery);
+    tox_options_set_start_port(log_options, 33445);
+    tox_options_set_end_port(log_options, 33445 + 2000);
+    tox_options_set_log_callback(log_options, &print_debug_log);
+    tox_options_set_log_user_data(log_options, log_user_data);
+    Tox *tox = tox_new(log_options, err);
+
+    if (options == nullptr) {
+        tox_options_free(log_options);
+    }
+
+    return tox;
+}
+
+Tox *tox_new_log(struct Tox_Options *options, Tox_Err_New *err, void *log_user_data)
+{
+    return tox_new_log_lan(options, err, log_user_data, false);
 }
