+++ auto_tests/conference_simple_test.c
@@ -3,6 +3,7 @@
 
 #include "../testing/misc_tools.h"
 #include "../toxcore/tox.h"
+#include "auto_test_support.h"
 #include "check_compat.h"
 
 typedef struct State {
