+++ auto_tests/conference_two_test.c
@@ -5,6 +5,7 @@
 
 #include "../testing/misc_tools.h"
 #include "../toxcore/tox.h"
+#include "auto_test_support.h"
 #include "check_compat.h"
 
 int main(void)
