+++ auto_tests/lan_discovery_test.c
@@ -3,6 +3,7 @@
 
 #include "../testing/misc_tools.h"
 #include "../toxcore/ccompat.h"
+#include "auto_test_support.h"
 
 int main(void)
 {
