+++ auto_tests/save_compatibility_test.c
@@ -2,6 +2,7 @@
 
 #include "../testing/misc_tools.h"
 #include "../toxcore/tox.h"
+#include "auto_test_support.h"
 #include "check_compat.h"
 
 #include <stdio.h>
