+++ auto_tests/save_friend_test.c
@@ -9,6 +9,7 @@
 #include "../toxcore/ccompat.h"
 #include "../toxcore/crypto_core.h"
 #include "../toxcore/tox.h"
+#include "auto_test_support.h"
 #include "check_compat.h"
 
 struct test_data {
