+++ auto_tests/set_status_message_test.c
@@ -10,6 +10,7 @@
 #include "../toxcore/ccompat.h"
 #include "../toxcore/tox.h"
 #include "../toxcore/util.h"
+#include "auto_test_support.h"
 #include "check_compat.h"
 
 #define STATUS_MESSAGE "Installing Gentoo"
