+++ auto_tests/tox_many_tcp_test.c
@@ -10,6 +10,7 @@
 #include "../toxcore/crypto_core.h"
 #include "../toxcore/tox.h"
 #include "../toxcore/util.h"
+#include "auto_test_support.h"
 #include "check_compat.h"
 
 /* The Travis-CI container responds poorly to ::1 as a localhost address
