+++ auto_tests/toxav_many_test.c
@@ -17,6 +17,7 @@
 #include "../toxcore/logger.h"
 #include "../toxcore/tox.h"
 #include "../toxcore/util.h"
+#include "auto_test_support.h"
 #include "check_compat.h"
 
 typedef struct {
